kconfig check: skip anbox check with pmaports.cfg (MR 2122)
Check for new variable supported_kconfig_anbox in pmaports.cfg, and do
not check for the anbox related kernel config options if this is not set
to True. This way we can disable the check for v21.06, where it is
currently failing, because CONFIG_PSI was added to
necessary_kconfig_options_anbox but the kernel configs were only
adjusted in edge.

We discussed whether to adjust the kernel configs on v21.06 instead, but
doing this rebuild of all affected kernels just to make CI happy didn't
seem like the right approach.