Add support for flashing to devices with Tow-Boot installed
https://tow-boot.org is a pre-configured u-boot which will be used on
the PPP to boot. Developer and Explorer Edition devices of this device
don't have a dedicated boot storage so Tow-Boot has to be installed on
the eMMC, on which the OS will also reside. Thus we should make sure we
don't overwrite an already installed Tow-Boot or we might risk bricking
people's devices

This uses https://sr.ht/~martijnbraam/findtow/ to detect if Tow-Boot is
installed on the specified storage