CONTRIBUTING: modernize (MR 2235)
Rewrite the contributing guide. Most of this was written a long time
ago, when pmaports and pmbootstrap still were in the same repository.
This time, focus on what is important for developing pmbootstrap and
omit some of the obvious information (how to make a merge request etc.).

Add relevent information from:
https://wiki.postmarketos.org/wiki/Pmbootstrap_development_guide

...so we can replace the wiki page with a link to this file. Having the
information in a CONTRIBUTING.md seems more intuitive than having it
somewhere in the wiki.

Mention that new or modified code should use f-strings, and that we
support all active python versions.